requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
pyright==1.1.384
base2048==0.1.3
